Rework types to be tagged unions
Originally, I had thought tagged unions were not ideal as I wanted a
typeid system. I had designed my initial Type structure to be able to
be packed into a 64-bit integer, but over time that goal was not
sufficient enough for my needs. Upon researching pattern matching, I
realized types can be stored in a global vector with indices
indicating the type id. So, this commit reworks types to be much
simpler and to support a type id system (without regards to modules yet).